chore: Rename module 'stratify_is_set' to 'stratify_is_set_warning'
For consistency with the other warning modules.
