+++ skore/src/skore/sklearn/train_test_split/warning/__init__.py
@@ -11,7 +11,7 @@ from .high_class_imbalance_warning import (
     HighClassImbalanceWarning,
 )
 from .random_state_unset_warning import RandomStateUnsetWarning
-from .stratify_is_set import StratifyWarning
+from .stratify_is_set_warning import StratifyWarning
 
 TRAIN_TEST_SPLIT_WARNINGS = [
     HighClassImbalanceTooFewExamplesWarning,
