+++ skore/src/skore/sklearn/train_test_split/warning/stratify_is_set_warning.py
